♻️ put restart service state in a service popup
